pyproject.toml
[tool.poetry]
name = "bewise"
version = "0.1.0"
description = ""
authors = ["Ayta Dzhivanov <kurt.alive.1994@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
psycopg2 = "^2.9.9"
alembic = "^1.12.0"
fastapi-sqlalchemy = "^0.2.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
